FIX: Bugfixed on line 408 np.zeros produces array with uint8 datatype So, another array can't use uint8 datatype format (calcarray) It needs float64!

+++ f2n/f2n/f2n.py
@@ -404,7 +404,8 @@ class f2nimage:
 			
 			
 			#calcarray.shape = numpyarrayshape
-			bwarray = np.zeros(calcarray.shape, dtype=np.uint8)
+			# bwarray = np.zeros(calcarray.shape, dtype=np.uint8)
+                        bwarray = np.zeros(calcarray.shape, dtype=np.float64)
 			calcarray.round(out=bwarray)
 			if negative:
 				if self.verbose:
